PurchaseNormStructureController.java - implemented everything except tracking user currency StructureForUser.java - x and y coordinates are floats not ints MobstersDbTables.java - constants for oil property in user table MobstersTableConstants.java - removed some old aoc2 constants TestUserEntityManager.java - relocated UserEntityManager EventStructure.proto - client dictates how much user spends upgrading struct
MonsterForUserService.java - created new comment section
MonsterForUserServiceImpl.java - above
QuestForUserService.java - above
QuestForUserServiceImpl.java - above
StructureForUserService.java - above; created new methods
UserService.java - above; deleted some aoc2 methods
UserSpellServiceImpl.java - commented out error

+++ mobsters-data-model/src/main/java/com/lvl6/mobsters/po/nonstaticdata/StructureForUser.java
@@ -32,10 +32,10 @@ public class StructureForUser extends BasePersistentObject<UUID>{
 	protected Date lastCollectTime = new Date();
 	
 	@Column(name="x_coordinate")
-	protected int xCoordinate = 0;
+	protected float xCoordinate = 0;
 	
 	@Column(name="y_coordinate")
-	protected int yCoordinate = 0;
+	protected float yCoordinate = 0;
 
 	@Column(name="purchase_time")
 	@Index
@@ -81,19 +81,19 @@ public class StructureForUser extends BasePersistentObject<UUID>{
 		this.lastCollectTime = lastCollectTime;
 	}
 
-	public int getxCoordinate() {
+	public float getxCoordinate() {
 		return xCoordinate;
 	}
 
-	public void setxCoordinate(int xCoordinate) {
+	public void setxCoordinate(float xCoordinate) {
 		this.xCoordinate = xCoordinate;
 	}
 
-	public int getyCoordinate() {
+	public float getyCoordinate() {
 		return yCoordinate;
 	}
 
-	public void setyCoordinate(int yCoordinate) {
+	public void setyCoordinate(float yCoordinate) {
 		this.yCoordinate = yCoordinate;
 	}
